A possible shortening of the paper so we fit in the page limit.
I hacked the bullet list in Section 3 so that it eats a lot less white space (but arguably doesn't look as nice). In the same section I did a little editing to shorten a few paragraphs by a line, and reduced the amount of white space below Table 1.

This gets us entirely within the page limit. With a little more editing we could possibly make room for the classification information?